tileEditor for v1.48
The color wheel stays open for every section by default.
On the new ‘Icon’ section, you can toggle between selecting an icon as before, or setting a background color for the icon itself (allowing one color for ‘on’ state, and another color for ‘off’ state if you want. This basically makes the tile background color a border if you use this feature. Otherwise you can still color the whole tile.)
When toggled, the background color in the edit section also changes, for clarity.